Fixed issue with scaleFactors
Issue #127
If scaleFactors are not read inside the same block, it can happen that the values are not in sync. Therefore implemented a warning for this case. Furthermore implemented parsing of scalingFactors before the parsing of the other values to ensure correct usage of scalingFactor even when the scaling factor is on a higher adress then the value to be calculated