Implement keep alive
For some modbus connections the topic of a disconnect after a specific timeout exists. In my case after 10s. This means if more than 10s no data is requested the connection will drop. To prevent this a keep alive is implemented by reading just single values which might even not exists. The interval can be sett in the config with the keepAliveInterval Parameter. If set to 0 the function is disabled